Refactor: split main.py into modular routers
Extract route handlers from main.py into dedicated router modules:
- ubiblio/routers/auth.py: authentication endpoints
- ubiblio/routers/books.py: book management endpoints
- ubiblio/routers/admin.py: admin endpoints
- ubiblio/routers/federation.py: federation endpoints
- ubiblio/routers/files.py: file handling endpoints
- ubiblio/routers/reading_lists.py: reading list endpoints
- ubiblio/dependencies.py: shared dependencies
